pyproject.toml
[project]
name = "microshop"
version = "0.1.0"
description = ""
authors = [
    {name = "eduard plyushchevich",email = "eduardplyushchevich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib[context] (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"

